Safari-609.1.11.1/LayoutTests/http/tests/xmlhttprequest/access-control-basic-denied-expected.txt

CONSOLE MESSAGE: line 24: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: line 24: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-denied.cgi due to access control checks.
PASS: Exception thrown. Cross-domain access was denied in 'send'. [ A network error occurred.].
